split ring implementations into modules, faster md5 ring rewrite
This work is related to issue #6 reported by @bjhockley

The main idea is to split the ketama implementation from the md5
or user configurable one to overcome some ketama logic problems.

This is required because ketama ring growth and shrink trigger a
full recalculation of the ring which is not efficient for rapidly
changing environments. The performance drawback is pretty bad.

The md5 (or customizable hash function) implementation now offers
a linear performance for growth and shrink operations inspired
from the openstack/tooz project.

IMPORTANT: this first step is meant to be backward compatibile.