Expose replicas argument for KetamaRing
The KetamaRing had a hardcoded value of 4 replicas. When migrating from
another library that uses a different replicacount uhashlib would not
produce the same values. To fix this we expose the replicas as a public
API and let the users decide how many replicas they want. This is for
example the case with the hash-lib pypi package, which uses 3 replicas.